changed model.py back to model_tiga_old and old model is model.py again
